requirements.txt
aiosqlite==0.21.0 ; python_version >= "3.12"
alembic==1.16.4 ; python_version >= "3.12"
annotated-types==0.7.0 ; python_version >= "3.12"
anyio==4.9.0 ; python_version >= "3.12"
asyncpg==0.30.0 ; python_version >= "3.12"
click==8.2.1 ; python_version >= "3.12"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.12"
databases==0.9.0 ; python_version >= "3.12"
fastapi==0.116.1 ; python_version >= "3.12"
greenlet==3.2.3 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
h11==0.16.0 ; python_version >= "3.12"
httptools==0.6.4 ; python_version >= "3.12"
idna==3.10 ; python_version >= "3.12"
mako==1.3.10 ; python_version >= "3.12"
markupsafe==3.0.2 ; python_version >= "3.12"
psycopg2-binary==2.9.10 ; python_version >= "3.12"
pydantic-core==2.33.2 ; python_version >= "3.12"
pydantic-settings==2.10.1 ; python_version >= "3.12"
pydantic==2.11.7 ; python_version >= "3.12"
pyjwt==2.10.1 ; python_version >= "3.12"
python-decouple==3.8 ; python_version >= "3.12"
python-dotenv==1.1.1 ; python_version >= "3.12"
pyyaml==6.0.2 ; python_version >= "3.12"
sniffio==1.3.1 ; python_version >= "3.12"
sqlalchemy==2.0.41 ; python_version >= "3.12"
starlette==0.47.2 ; python_version >= "3.12"
typing-extensions==4.14.1 ; python_version >= "3.12"
typing-inspection==0.4.1 ; python_version >= "3.12"
uvicorn==0.35.0 ; python_version >= "3.12"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.12"
watchfiles==1.1.0 ; python_version >= "3.12"
websockets==15.0.1 ; python_version >= "3.12"
